Lyndon Browser: bring the Windows build up to the Linux one
The Linux side gained password import and export, hand editing, a save
bar that asks for whichever half of a login the page would not give up,
and windows that reopen at the size they were left. None of that reached
Windows, and two of the shared files had drifted apart.

Everything below the toolkit is now genuinely one copy. import.c and
pwfile.c compile for both, and the credential struct and the origin rule
that autofill matches on moved into credentials.c so that a saved login
cannot mean two different things on two platforms. Chromium's key is
found through DPAPI and its values decrypted with AES-256-GCM through
CNG, next to the Secret Service and OpenSSL path, with the walk over the
profile's table written once; Chrome 127's app-bound rows are counted as
unreadable rather than guessed at. Firefox goes through NSS on both,
detected the same way.

New on Windows: win32/login.c, the save-login card, as a popup because
WebView2 owns the whole page area; the import page now chooses between
bookmarks, history and saved logins; the passwords page gained an editor,
a file import, a CSV export and a copy button. ly_passwords_list now
loads secrets there as libsecret already did, which is what an export
needs. The message loop drains the GLib context, so the debounced config
save fires on a timer rather than only at exit.

+++ Lyndon Browser/src/pwfile.c
@@ -416,7 +416,7 @@ sheet_end (GMarkupParseContext *ctx, const char *element,
 
     if (parse->is_shared) {
       gint64 index = g_ascii_strtoll (text, NULL, 10);
-      resolved = (index >= 0 && index < parse->shared->len)
+      resolved = (index >= 0 && index < (gint64) parse->shared->len)
         ? g_strdup (g_ptr_array_index (parse->shared, index))
         : g_strdup ("");
     } else {
